feat: split components
