+++ frontend/src/scenes/llm-observability/LLMObservabilityTraces.tsx
@@ -1,0 +1,3 @@
+export function LLMObservabilityTraces(): JSX.Element {
+    return <div>Traces</div>
+}
